refactor: 서버-클라이언트 아키텍처 분리 및 Ktor 기반 API 구축
**모듈 구조 변경:**
- `server`: Ktor 기반 REST API 서버 (포트 8080)
- `shared`: 공통 데이터 모델 및 유틸리티 (JVM/JS 멀티플랫폼)
- `site`: Kobweb 프론트엔드 전용 (포트 8081)

**주요 변경사항:**
- HTTP 클라이언트 도입: `ApiClient`를 통한 서버-클라이언트 통신
- 데이터 로딩 방식 변경: localStorage → 서버 API 호출
- 패키지 구조 통일: `net.lateinit.bug_or_feature.*`
- `site` 모듈 jvmMain 제거 (프론트엔드 단일 책임)

**기술 스택:**
- 서버: Ktor + Exposed + SQLite
- 클라이언트: Ktor HTTP Client + Kobweb
- 공통: Kotlinx Serialization

**API 엔드포인트:**
- `GET /api/prompts`: 질문 목록 조회
- `POST /api/prompts`: 새 질문 추가
- `POST /api/prompts/{id}/vote`: 투표 처리
