+++ shared/src/commonMain/kotlin/net/lateinit/bug_or_feature/shared/util/TimeUtil.kt
@@ -1,0 +1,3 @@
+package net.lateinit.bug_or_feature.shared.util
+
+expect fun nowMillis(): Long
