+++ shared/src/jsMain/kotlin/net/lateinit/bug_or_feature/shared/util/TimeUtil.js.kt
@@ -1,0 +1,3 @@
+package net.lateinit.bug_or_feature.shared.util
+
+actual fun nowMillis(): Long = (js("Date.now()") as Double).toLong()
