+++ shared/src/jvmMain/kotlin/net/lateinit/bug_or_feature/shared/util/TimeUtil.jvm.kt
@@ -1,0 +1,3 @@
+package net.lateinit.bug_or_feature.shared.util
+
+actual fun nowMillis(): Long = System.currentTimeMillis()
