--- site/src/jsMain/kotlin/net/lateinit/bug_or_feature/AppEntry.kt
@@ -1,29 +1,0 @@
-package net.lateinit.bug_or_feature
-
-import androidx.compose.runtime.*
-import com.varabyte.kobweb.compose.ui.Modifier
-import com.varabyte.kobweb.compose.ui.modifiers.*
-import com.varabyte.kobweb.core.App
-import com.varabyte.kobweb.silk.SilkApp
-import com.varabyte.kobweb.silk.components.layout.Surface
-import com.varabyte.kobweb.silk.init.InitSilk
-import com.varabyte.kobweb.silk.init.InitSilkContext
-import com.varabyte.kobweb.silk.init.registerStyleBase
-import com.varabyte.kobweb.silk.style.common.SmoothColorStyle
-import com.varabyte.kobweb.silk.style.toModifier
-import org.jetbrains.compose.web.css.*
-
-@InitSilk
-fun initStyles(ctx: InitSilkContext) {
-    ctx.stylesheet.registerStyleBase("html, body") { Modifier.fillMaxHeight() }
-}
-
-@App
-@Composable
-fun AppEntry(content: @Composable () -> Unit) {
-    SilkApp {
-        Surface(SmoothColorStyle.toModifier().fillMaxHeight()) {
-            content()
-        }
-    }
-}
