--- site/src/jsMain/kotlin/net/lateinit/bug_or_feature/util/Share.kt
@@ -1,10 +1,0 @@
-package net.lateinit.bug_or_feature.util
-
-import kotlinx.browser.window
-
-fun shareLink(id: String?) {
-    if (id != null) {
-        window.location.hash = "#q=$id"
-    }
-}
-
