--- site/src/jvmMain/kotlin/net/lateinit/bug_or_feature/api/README.md
@@ -1,3 +1,0 @@
-Define API routes in here.
-
-See also: https://kobweb.varabyte.com/docs/concepts/server/fullstack#define-api-routes
